feat: Implement GetPrettyName for macOS, Linux, and Windows (#96)
* feat: implement GetPrettyName for macOS, Linux, and Windows


* chore(deps): Make golang.org/x/sys dependency direct


---------
